+++ src/main/java/com/onlineshop/catalog/api/events/ItemRefiledEvent.java
@@ -11,11 +11,11 @@ import static com.onlineshop.catalog.api.events.ItemEventValue.ITEM_REFILED;
 
 @DomainEvent(name = ITEM_REFILED)
 public class ItemRefiledEvent extends Event<CatalogAggregate> {
-    public ItemRefiledEvent(UUID id, int amount) {
+    public ItemRefiledEvent(UUID id, long amount) {
         super(UUID.randomUUID(), ITEM_REFILED, 1, new Date().getTime());
         this.id = id;
         this.amount = amount;
     }
     final UUID id;
-    final int amount;
+    final long amount;
 }
