+++ src/main/java/com/onlineshop/catalog/dto/ItemCreateDto.java
@@ -3,7 +3,7 @@ package com.onlineshop.catalog.dto;
 public record ItemCreateDto(
     String name,
     String description,
-    double price,
-    double amount
+    long price,
+    long amount
 ){
 }
